+++ test/bai.test.ts
@@ -1,6 +1,6 @@
 import { BAI, BamFile, BamRecord } from '../src'
 
-import { LocalFile } from 'generic-filehandle'
+import { LocalFile } from 'generic-filehandle2'
 import FakeRecord from './fakerecord'
 
 test('loads BAI volvox-sorted.bam.bai', async () => {